agent/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
nanoid==2.0.0
openai>=1.60.0
azure-identity>=1.15.0
github-copilot-sdk>=0.1.21
